Create AddEditUser component for adding and editing modals
